// packages/nuxt/src/generators/application/schema.d.ts
import { Linter } from '@nx/linter';

export interface Schema {
  name: string;
  tags?: string;
  directory?: string;
  unitTestRunner?: 'vitest' | 'none';
  e2eTestRunner?: 'cypress' | 'playwright' | 'none';
  inSourceTests?: boolean;
  linter: Linter;
  skipFormat?: boolean;
  rootProject?: boolean;
  skipNxJson?: boolean;
}
export interface NormalizedSchema<T extends Schema = Schema> extends T {
  appProjectRoot: string;
  appProjectName: string;
  e2eProjectName: string;
  e2eProjectRoot: string;
  projectDirectory: string;
  parsedTags: string[];
  unitTestRunner: 'vitest' | 'none';
  e2eTestRunner: 'cypress' | 'playwright' | 'none';
  rootProject: boolean;
}